requirements.txt
Flask==2.1.3
Flask-Cors  # CORS 허용 (필요한 경우)
requests
werkzeug==2.0.3
celery[redis]  # Celery + Redis 지원
redis
Pillow
gunicorn